requirements.txt
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
boto3==1.20.45; python_version >= "3.6"
botocore==1.23.45; python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3"
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
cycler==0.11.0; python_version >= "3.7"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
filelock==3.4.2; python_version >= "3.7"
fonttools==4.29.1; python_version >= "3.7"
gdown==4.2.0
grpcio==1.43.0; python_version >= "3.6"
higher==0.2.1; python_full_version >= "3.6.0"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-resources==5.4.0; python_version < "3.9" and python_version >= "3.7"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
joblib==1.1.0; python_version >= "3.7"
jsonschema==4.4.0; python_version >= "3.7"
kiwisolver==1.3.2; python_version >= "3.7"
llvmlite==0.34.0; python_version >= "3.6"
matplotlib==3.5.1; python_version >= "3.7"
msgpack==1.0.3
numba==0.51.2; python_version >= "3.6"
numpy==1.22.1; python_version >= "3.8"
packaging==21.3; python_version >= "3.7"
pandas==1.3.5; python_full_version >= "3.7.1"
pillow==9.0.0; python_version >= "3.7"
protobuf==3.19.3; python_version >= "3.5"
pynndescent==0.5.6
pyparsing==3.0.6; python_version >= "3.7"
pyrsistent==0.18.1; python_version >= "3.7"
pysocks==1.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
python-dateutil==2.8.2; python_full_version >= "3.7.1" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7")
pytz==2021.3; python_full_version >= "3.7.1"
pyyaml==6.0; python_version >= "3.6"
qpsolvers==1.8.0; python_version >= "3.6"
quadprog==0.1.11; python_version >= "3.6"
ray==1.10.0
redis==4.1.1; python_version >= "3.6"
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
s3transfer==0.5.0; python_version >= "3.6"
scikit-learn==1.0.2; python_version >= "3.7"
scipy==1.7.3; python_version >= "3.7" and python_version < "3.11"
setuptools-scm==6.4.2; python_version >= "3.7"
six==1.16.0; python_full_version >= "3.7.1" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7")
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
tabulate==0.8.9
tensorboardx==2.4.1
threadpoolctl==3.1.0; python_version >= "3.7"
tomli==2.0.1; python_version >= "3.7"
torch==1.10.1; python_full_version >= "3.6.2"
torchvision==0.11.2
tqdm==4.62.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.2"
umap-learn==0.5.2
urllib3==1.26.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
wrapt==1.13.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
zipp==3.7.0; python_version < "3.9" and python_version >= "3.7"
